Move pageAnchor to src/app/lib/utilities
