+++ src/app/containers/Image/index.jsx
@@ -9,7 +9,7 @@ import {
 } from '../../lib/styledGrid';
 import createSrcset from './helpers/srcSet';
 import getIChefURL from './helpers/ichefUrl';
-import urlWithPageAnchor from '../../helpers/pageAnchor';
+import urlWithPageAnchor from '../../lib/utilities/pageAnchor';
 
 const DEFAULT_IMAGE_RES = 640;
 const LAZYLOAD_FROM_BLOCK = 3;
